reorganize to torch directories
